+++ ts/app/prog.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
 
-from ts.model.univariate.deep import RnnForecast
+from ts.model.univariate.multiseq.deep import RnnForecast
 from ts.data.univariate.nonexo import StandardGenerator
 from ts.log import GlobalLogger
