+++ ts/app/univ_extreme_time2.py
@@ -1,9 +1,8 @@
-import os
 import numpy as np
 import tensorflow as tf
 
 from ts.data.univariate.nonexo import StandardGenerator
-from ts.model.univariate.deep import ExtremeTime2
+from ts.model.univariate.oneseq.deep import ExtremeTime2
 from ts.plot import Plot
