--- ts/model/univariate/deep/__init__.py
@@ -1,6 +1,0 @@
-from ts.model.univariate.deep.extreme_time import ExtremeTime
-from ts.model.univariate.deep.extreme_time2 import ExtremeTime2
-from ts.model.univariate.deep.rnn_forecast import RnnForecast
-from ts.model.univariate.deep.simple_rnn_forecast import SimpleRnnForecast
-from ts.model.univariate.deep.gru_forecast import GruForecast
-from ts.model.univariate.deep.lstm_forecast import LstmForecast
