+++ ts/model/univariate/multiseq/deep/__init__.py
@@ -1,0 +1,4 @@
+from ts.model.univariate.multiseq.deep.rnn_forecast import RnnForecast
+from ts.model.univariate.multiseq.deep.simple_rnn_forecast import SimpleRnnForecast
+from ts.model.univariate.multiseq.deep.gru_forecast import GruForecast
+from ts.model.univariate.multiseq.deep.lstm_forecast import LstmForecast
