+++ ts/model/univariate/oneseq/deep/__init__.py
@@ -1,0 +1,2 @@
+from ts.model.univariate.oneseq.deep.extreme_time import ExtremeTime
+from ts.model.univariate.oneseq.deep.extreme_time2 import ExtremeTime2
